fix: fix balance calculation and form validation
